Code iteration study (#1)
* Measure calib4 iteration time and background in an automated fashion

* Measure wrtsig() iteration time and background in an automated fashion

* Put the set/get timing pair at the right position, misplaced them in previous commit

* Measure the impact of the number of iterations of mrst4

* more wrtsig studies

* More mrst4 studies

* Use delay_us function between mrst4 and wrtsig

To get a clearer image of what happens with the number of active pixels
as a function of delay between finishing the calibration and starting
the wrtsig waveform I implemented the delay_us between these waveforms

* delay between wrstsig and readout

* delay between mrst4 and readout, wrtsig outcommented

* Idle waveform between mrst4 and readout, no wrtsig

* idle waveform iteration between mrst4=1 and wrtsig=1

* Do 10x iterions of idle between wrtsig and mrst4

* paralelized calin4

* The code I put previously into calin4 was supposed to be in calib4. I fixed that now.

* Fixed codeIterationLimit initialization

codeIterationLimit was first initialized to 0, later if it is bigger the
sit limit it is set to 1. If the code iteration limits are bigger than 1
this will lead to a set of measurements for a codeIteration number that
is only repeated once. So I fixed it.

* Changed teh delay in calib4 to 5 cycles (from 1).

I wanna see if this makes any difference. 5 cycles if approximately how
long it takes to outpur a pin word.

* Removed the idle between wrtsig and readout

This should not have been here. I should have removed it a long time
ago. Despite it's presence it should not have affected the results.

* Setup for code iteration study of wrtsig.

Timer measures now time between ending of calibration and and of wrtsig.

* parallelized the remaining waveforms

Calin4
idle4
mrst4
wrtsig

drdtst is not suited for parallelization/does not benefit from it, just
changed some delays and use parallelization just at the beginning to set
the initial pins

* campaign 67, wrtsig study

revert to non parallel wrtsig

* Run at 25MHz and the final configuration of wrtsig that I wanna use